requirements.txt
opencv-python-headless
imgbeddings
psycopg2-binary
grpcio-tools
python-dotenv
lgpio
mfrc522
RPi.GPIO
rpi-lgpio


comando generador de archivos para grpc
python -m grpc_tools.protoc -I. --python_out=. --pyi_out=. --grpc_python_out=. ./access.proto